modification sur chaque controleur

+++ EasyColoc/app/Http/Controllers/CategoryController.php
@@ -3,68 +3,34 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Colocation;
 use App\Http\Requests\StoreCategoryRequest;
-use App\Http\Requests\UpdateCategoryRequest;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        $categories = Category::all();
-        return view('categories.index', compact('categories'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        return view('categories.create');
-    }
-
     /**
      * Store a newly created resource in storage.
      */
     public function store(StoreCategoryRequest $request)
     {
+        $colocation = Colocation::findOrFail($request->colocation_id);
+        
+       
+        $isOwner = $colocation->users()
+            ->where('user_id', auth()->id())
+            ->wherePivot('role', 'owner')
+            ->exists();
+
+        if (!$isOwner) {
+            return redirect()->back()->with('error', 'Seul le propriétaire peut gérer les catégories.');
+        }
+
         Category::create([
             'title' => $request->title,
             'colocation_id' => $request->colocation_id,
         ]);
-        return redirect()->route('categories.index')
+        return redirect()->back()
             ->with('success', 'Category created successfully.');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Category $category)
-    {
-        return view('categories.show', compact('category'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Category $category)
-    {
-        return view('categories.edit', compact('category'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateCategoryRequest $request, Category $category)
-    {
-        $category->update([
-            'title' => $request->title,
-            'colocation_id' => $request->colocation_id,
-        ]);
-        return redirect()->route('categories.index')
-        ->with('success', 'Category updated successfully.');
     }
 
     /**
@@ -72,8 +38,19 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $colocation = $category->colocation;
+        
+        $isOwner = $colocation->users()
+            ->where('user_id', auth()->id())
+            ->wherePivot('role', 'owner')
+            ->exists();
+
+        if (!$isOwner) {
+            return redirect()->back()->with('error', 'Seul le propriétaire peut gérer les catégories.');
+        }
+
         $category->delete();
-            return redirect()->route('categories.index')
+            return redirect()->back()
             ->with('success', 'Category deleted successfully.');
     }
 }
